+++ configs.py
@@ -1,0 +1,4 @@
+import yaml
+
+file = open("resource/app.conf.yaml")
+AppConfig = yaml.load(file, Loader=yaml.FullLoader)
